Refactor JsonApiDiscriminators with base impl due to a lot of duplication and imporvements
