+++ core/src/main/java/com/infinum/jsonapix/core/common/JsonApiXMissingArgumentException.kt
@@ -2,6 +2,7 @@ package com.infinum.jsonapix.core.common
 
 import java.lang.IllegalArgumentException
 
+// todo NIKOLA add attribute logging
 class JsonApiXMissingArgumentException : IllegalArgumentException() {
     override val message: String = "Missing attribute"
 }
