+++ core/src/main/java/com/infinum/jsonapix/core/discriminators/JsonApiDiscriminator.kt
@@ -21,14 +21,13 @@ import kotlinx.serialization.json.jsonObject
  * Child objects -> Child prefix + type parameter e.g. Attributes_person where person is the type
  * of a class called Person passed as a parameter to JsonApiX annotation.
  */
-@SuppressWarnings("TooManyFunctions")
 class JsonApiDiscriminator(
     private val rootType: String,
     private val rootLinks: String,
     private val resourceObjectLinks: String,
     private val relationshipsLinks: String,
     private val meta: String
-) : Discriminator {
+) : BaseJsonApiDiscriminator(relationshipsLinks, meta) {
 
     private val rootDiscriminator = CommonDiscriminator(rootType)
 
@@ -146,37 +145,11 @@ class JsonApiDiscriminator(
         }
     }
 
-    private fun getJsonObjectEntry(key: String, data: JsonElement): Map.Entry<String, JsonElement> {
-        return object : Map.Entry<String, JsonElement> {
-            override val key: String = key
-            override val value: JsonElement = data
-        }
-    }
-
-    private fun getJsonArrayEntry(key: String, data: JsonArray): Map.Entry<String, JsonArray> {
-        return object : Map.Entry<String, JsonArray> {
-            override val key: String = key
-            override val value: JsonArray = data
-        }
-    }
-
-    private fun getDataObject(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.DATA]
-
-    private fun getRelationshipsObject(jsonElement: JsonElement) =
+    override fun getRelationshipsObject(jsonElement: JsonElement): JsonElement? =
         getDataObject(jsonElement)?.jsonObject?.get(JsonApiConstants.Keys.RELATIONSHIPS)
 
-    private fun getAttributesObject(jsonElement: JsonElement) =
+    override fun getAttributesObject(jsonElement: JsonElement): JsonElement? =
         getDataObject(jsonElement)?.jsonObject?.get(JsonApiConstants.Keys.ATTRIBUTES)
-
-    private fun getIncludedArray(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.INCLUDED]
-
-    private fun getLinksObject(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.LINKS]
-
-    private fun getMetaObject(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.META]
 
     private fun getJsonObjectWithDataDiscriminator(
         original: JsonElement,
@@ -210,57 +183,5 @@ class JsonApiDiscriminator(
             resultMap.putAll(entries.map { Pair(it.key, it.value) })
             JsonObject(resultMap)
         }
-    }
-
-    private fun getNewDataObject(
-        original: JsonElement,
-        attributesObject: JsonElement?,
-        relationshipsObject: JsonElement?,
-        linksObject: JsonElement?
-    ): JsonObject {
-        return original.jsonObject.entries.toMutableSet().let { entries ->
-            attributesObject?.let { attributes ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.ATTRIBUTES }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.ATTRIBUTES, attributes))
-            }
-
-            relationshipsObject?.let { relationships ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.RELATIONSHIPS }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.RELATIONSHIPS, relationships))
-            }
-
-            linksObject?.let { links ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.LINKS }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.LINKS, links))
-            }
-
-            val resultMap = mutableMapOf<String, JsonElement>()
-            resultMap.putAll(entries.map { Pair(it.key, it.value) })
-            JsonObject(resultMap)
-        }
-    }
-
-    private fun getNewRelationshipsObject(
-        original: JsonElement
-    ): JsonObject {
-        val resultMap = mutableMapOf<String, JsonElement>()
-        val relationshipsLinksDiscriminator = CommonDiscriminator(relationshipsLinks)
-        original.jsonObject.entries.filter { it.value is JsonObject }.forEach { relationshipEntry ->
-            val set = relationshipEntry.value.jsonObject.entries.toMutableSet()
-            getLinksObject(relationshipEntry.value)?.let { linksSafe ->
-                val newLinks = relationshipsLinksDiscriminator.inject(linksSafe)
-                set.removeAll { it.key == JsonApiConstants.Keys.LINKS }
-                set.add(getJsonObjectEntry(JsonApiConstants.Keys.LINKS, newLinks))
-            }
-            val tempMap = mutableMapOf<String, JsonElement>()
-            tempMap.putAll(set.map { Pair(it.key, it.value) })
-            resultMap[relationshipEntry.key] = JsonObject(tempMap)
-        }
-        return JsonObject(resultMap)
-    }
-
-    private fun getNewMetaObject(original: JsonElement): JsonElement {
-        val metaDiscriminator = CommonDiscriminator(meta)
-        return metaDiscriminator.inject(original)
     }
 }
