+++ core/src/main/java/com/infinum/jsonapix/core/discriminators/JsonApiListDiscriminator.kt
@@ -9,14 +9,13 @@ import kotlinx.serialization.json.buildJsonArray
 import kotlinx.serialization.json.jsonArray
 import kotlinx.serialization.json.jsonObject
 
-@SuppressWarnings("TooManyFunctions")
 class JsonApiListDiscriminator(
     private val rootType: String,
     private val rootLinks: String,
     private val resourceObjectLinks: String,
     private val relationshipsLinks: String,
     private val meta: String
-) : Discriminator {
+) : BaseJsonApiDiscriminator(relationshipsLinks, meta) {
 
     private val rootDiscriminator = CommonDiscriminator(rootType)
 
@@ -24,7 +23,7 @@ class JsonApiListDiscriminator(
     @SuppressWarnings("SwallowedException", "TooGenericExceptionCaught", "LongMethod")
     override fun inject(jsonElement: JsonElement): JsonElement {
         try {
-            val dataArray = getDataArray(jsonElement)
+            val dataArray = getDataObject(jsonElement)
             val includedObject = getIncludedArray(jsonElement)
             val newDataEntries = mutableListOf<JsonElement>()
             val rootLinksObject = getLinksObject(jsonElement)
@@ -115,7 +114,7 @@ class JsonApiListDiscriminator(
     @SuppressWarnings("SwallowedException", "TooGenericExceptionCaught")
     override fun extract(jsonElement: JsonElement): JsonElement {
         try {
-            val dataArray = getDataArray(jsonElement)?.let {
+            val dataArray = getDataObject(jsonElement)?.let {
                 rootDiscriminator.extract(it)
             }
             val includedArray = getIncludedArray(jsonElement)?.let { included ->
@@ -144,37 +143,11 @@ class JsonApiListDiscriminator(
         }
     }
 
-    private fun getJsonObjectEntry(key: String, data: JsonElement): Map.Entry<String, JsonElement> {
-        return object : Map.Entry<String, JsonElement> {
-            override val key: String = key
-            override val value: JsonElement = data
-        }
-    }
-
-    private fun getJsonArrayEntry(key: String, data: JsonArray): Map.Entry<String, JsonArray> {
-        return object : Map.Entry<String, JsonArray> {
-            override val key: String = key
-            override val value: JsonArray = data
-        }
-    }
-
-    private fun getDataArray(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.DATA]
-
-    private fun getRelationshipsObject(jsonElement: JsonElement) =
+    override fun getRelationshipsObject(jsonElement: JsonElement): JsonElement? =
         jsonElement.jsonObject[JsonApiConstants.Keys.RELATIONSHIPS]
 
-    private fun getAttributesObject(jsonElement: JsonElement) =
+    override fun getAttributesObject(jsonElement: JsonElement): JsonElement? =
         jsonElement.jsonObject[JsonApiConstants.Keys.ATTRIBUTES]
-
-    private fun getIncludedArray(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.INCLUDED]
-
-    private fun getLinksObject(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.LINKS]
-
-    private fun getMetaObject(jsonElement: JsonElement) =
-        jsonElement.jsonObject[JsonApiConstants.Keys.META]
 
     private fun getJsonObjectWithDataDiscriminator(
         original: JsonElement,
@@ -208,57 +181,5 @@ class JsonApiListDiscriminator(
             resultMap.putAll(entries.map { Pair(it.key, it.value) })
             JsonObject(resultMap)
         }
-    }
-
-    private fun getNewDataObject(
-        original: JsonElement,
-        attributesObject: JsonElement?,
-        relationshipsObject: JsonElement?,
-        linksObject: JsonElement?
-    ): JsonObject {
-        return original.jsonObject.entries.toMutableSet().let { entries ->
-            attributesObject?.let { attributes ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.ATTRIBUTES }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.ATTRIBUTES, attributes))
-            }
-
-            relationshipsObject?.let { relationships ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.RELATIONSHIPS }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.RELATIONSHIPS, relationships))
-            }
-
-            linksObject?.let { links ->
-                entries.removeAll { it.key == JsonApiConstants.Keys.LINKS }
-                entries.add(getJsonObjectEntry(JsonApiConstants.Keys.LINKS, links))
-            }
-
-            val resultMap = mutableMapOf<String, JsonElement>()
-            resultMap.putAll(entries.map { Pair(it.key, it.value) })
-            JsonObject(resultMap)
-        }
-    }
-
-    private fun getNewRelationshipsObject(
-        original: JsonElement
-    ): JsonObject {
-        val resultMap = mutableMapOf<String, JsonElement>()
-        val relationshipsLinksDiscriminator = CommonDiscriminator(relationshipsLinks)
-        original.jsonObject.entries.filter { it.value is JsonObject }.forEach { relationshipEntry ->
-            val set = relationshipEntry.value.jsonObject.entries.toMutableSet()
-            getLinksObject(relationshipEntry.value)?.let { linksSafe ->
-                val newLinks = relationshipsLinksDiscriminator.inject(linksSafe)
-                set.removeAll { it.key == JsonApiConstants.Keys.LINKS }
-                set.add(getJsonObjectEntry(JsonApiConstants.Keys.LINKS, newLinks))
-            }
-            val tempMap = mutableMapOf<String, JsonElement>()
-            tempMap.putAll(set.map { Pair(it.key, it.value) })
-            resultMap[relationshipEntry.key] = JsonObject(tempMap)
-        }
-        return JsonObject(resultMap)
-    }
-
-    private fun getNewMetaObject(original: JsonElement): JsonElement {
-        val metaDiscriminator = CommonDiscriminator(meta)
-        return metaDiscriminator.inject(original)
     }
 }
